backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.13.0
yt-dlp==2025.3.31
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.1
pydantic-settings==2.8.1
